Move buildSongListFromIds method to MediaStoreUtil
