+++ app/src/main/java/com/marverenic/music/data/store/MediaStoreUtil.java
@@ -734,4 +734,83 @@ public final class MediaStoreUtil {
 
         return 0;
     }
+
+    /**
+     * Build an {@link ArrayList} of {@link Song}s from a list of id's. Doesn't require the
+     * library to be loaded
+     * @param songIDs The list of song ids to convert to {@link Song}s
+     * @param context The {@link Context} used to open a {@link Cursor}
+     * @return An {@link ArrayList} of {@link Song}s with ids matching those of the
+     *         songIDs parameter
+     */
+    public static List<Song> buildSongListFromIds(long[] songIDs, Context context) {
+        List<Song> contents = new ArrayList<>();
+        // Split this request into batches of size SQL_MAX_VARS
+        for (int i = 0; i < songIDs.length / SQL_MAX_VARS; i++) {
+            contents.addAll(buildSongListFromIds(songIDs, context,
+                    i * SQL_MAX_VARS, (i + 1) * SQL_MAX_VARS));
+        }
+
+        // Load the remaining songs (the last section that's not divisible by SQL_MAX_VARS)
+        contents.addAll(buildSongListFromIds(songIDs, context,
+                SQL_MAX_VARS * (songIDs.length / SQL_MAX_VARS), songIDs.length));
+
+        // Sort the contents of the list so that it matches the order of the array
+        List<Song> songs = new ArrayList<>();
+        for (long i : songIDs) {
+            for (Song s : contents) {
+                if (s.getSongId() == i) {
+                    songs.add(s);
+                    break;
+                }
+            }
+        }
+
+        return songs;
+    }
+
+    /**
+     * Implementation of {@link MediaStoreUtil#buildSongListFromIds(long[], Context)}. This method
+     * takes upper and lower bounds into account when looking at the song ids so that it can be
+     * partitioned into sections of size {@link #SQL_MAX_VARS} without the overhead of making array
+     * copies.
+     * @param songIDs The song ids build the list from
+     * @param context A Context to open a {@link Cursor} to query the {@link MediaStore}
+     * @param lowerBound The first index in the array to get IDs from
+     * @param upperBound The last index in the array to get IDs from
+     * @return An unsorted list of {@link Song Songs} with the same IDs as the ids that were passed
+     * into {@code songIDs}
+     */
+    private static List<Song> buildSongListFromIds(long[] songIDs, Context context, int lowerBound,
+                                                   int upperBound) {
+        List<Song> contents = new ArrayList<>();
+        if (songIDs.length == 0) {
+            return contents;
+        }
+
+        String query = MediaStore.Audio.Media._ID + " IN(?";
+        String[] ids = new String[upperBound - lowerBound];
+        ids[0] = Long.toString(songIDs[lowerBound]);
+
+        for (int i = 1; i < ids.length; i++) {
+            query += ",?";
+            ids[i] = Long.toString(songIDs[i + lowerBound]);
+        }
+        query += ")";
+
+        Cursor cur = context.getContentResolver().query(
+                MediaStore.Audio.Media.EXTERNAL_CONTENT_URI,
+                SONG_PROJECTION,
+                query, ids,
+                MediaStore.Audio.Media.TITLE + " ASC");
+
+        if (cur == null) {
+            return contents;
+        }
+
+        contents = Song.buildSongList(cur, context.getResources());
+        cur.close();
+
+        return contents;
+    }
 }
